Single entry point refactor, multiprocessing
Now all analysis functions can be launched from one driver function.
This is also enables multiprocessing. So far, this has been implemented
for the embedding, but it can easily be added for others (and a bunch of
code can be removed).

In doing this refactor, I also centralized dataset information. This
simplifies access to most of the day that we're most interested in for a
given dataset (and makes comparison between datasets easier)