Refactor filling calculated values, add embedded hists
I added true and hybrid hists for the embedding because they need a
special set of cuts (at least the double counting cut) that doesn't
apply to the standard data analysis. So I either had to implement an
independent cuts function and add more branches that the data doesn't
care about, or combine the functionality into the embedded tree
analysis. I decided on the later.

This also added some kt cuts to the matching, but in doing so, the
filling of values finally hit the breaking point of being too messy, so
I developed a technique to reduce repeated code. Now, each method is
calculated and filled in a single function (taking advantage of a
unbound method and partial).