Fix subjet edges to splitting edges
In some cases, we were assigning subjet edges to the wrong splitting
edges (specifically seen when drawing a recursive splitting near the end
of the iterative chain). In those cases, it would assign an iterative
subjet as if source of a recursive node. This caused quite some concern,
but in the end it turned out just to be a bug in drawing the splittings
(not in the splittings themselves)