Standardize bailing out early, bug fixes
So parsl doesn't pay attention to input and output files... It appeared
as if it did because I had an early bailout for one function if the file
already existed. So we add it manually. This could be handled better,
but for now we do the dumb thing and just check in the app.

Also fixed a bug in determining the output filename