Explore setting branch status to save AOD read time
Due to the AOD design, it wasn't fully split, so we have limited control
of which branches are enabled via the standard methods. I also tried
manual methods, and they appear as if they should work.  Unfortunately,
the status reported by ROOT isn't really reliable in terms of whether
the branch is actually disabled. So we took the straightforward savings
of ~ 25-30%, and moved on